feat(PageHeader): add empty state
