+++ packages/react/src/experimental/Navigation/Header/PageHeader/index.tsx
@@ -17,17 +17,8 @@ import { Tooltip } from "../../../Overlays/Tooltip"
 import { useSidebar } from "../../ApplicationFrame/FrameProvider"
 import { Dropdown } from "../../Dropdown"
 
-import ChevronRight from "@/icons/app/ChevronRight.tsx"
-import Messages from "@/icons/app/Messages.tsx"
-import { Image } from "@/lib/imageHandler"
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/ui/dropdown-menu.tsx"
 import Breadcrumbs, { type BreadcrumbItemType } from "../Breadcrumbs"
+import { ProductUpdates, ProductUpdatesProp } from "../ProductUpdates"
 
 export type PageAction = {
   label: string
@@ -73,11 +64,7 @@ type HeaderProps = {
   breadcrumbs?: BreadcrumbItemType[]
   productUpdates?: {
     isVisible?: boolean
-    label: string
-    updatesPageUrl: UrlString
-    getUpdatesQuery: () => Promise<Array<ProductUpdate>>
-    hasUnread?: boolean
-  }
+  } & ProductUpdatesProp
 }
 
 function PageNavigationLink({
@@ -255,12 +242,7 @@ export function PageHeader({
         )}
         {hasProductUpdates && (
           <div className="items-right flex gap-2">
-            <ProductUpdates
-              updatesPageUrl={productUpdates?.updatesPageUrl}
-              getUpdatesQuery={productUpdates?.getUpdatesQuery}
-              hasUnread={productUpdates?.hasUnread}
-              label="Product updates"
-            />
+            <ProductUpdates {...productUpdates} />
           </div>
         )}
         {hasActions && (
@@ -297,159 +279,5 @@ function PageAction({ action }: { action: PageAction }): ReactElement {
     >
       <Icon icon={action.icon} size="md" />
     </Link>
-  )
-}
-
-type ProductUpdate = {
-  title: string
-  href: UrlString
-  imageURL: UrlString
-  updated: string
-  unread?: boolean
-}
-
-type ProductUpdatesProp = {
-  label: string
-  updatesPageUrl: UrlString
-  getUpdatesQuery: () => Promise<Array<ProductUpdate>>
-  hasUnread?: boolean
-}
-const ProductUpdates = ({
-  label,
-  getUpdatesQuery,
-  updatesPageUrl,
-  // hasUnread = false,
-}: ProductUpdatesProp) => {
-  const [updates, setUpdates] = useState<Array<ProductUpdate> | null>(null)
-  const [featuredUpdate, ...restUpdates] = updates ?? []
-  return (
-    <DropdownMenu
-      onOpenChange={async (open) => {
-        if (open && updates === null) {
-          const response = await getUpdatesQuery()
-          setUpdates(response)
-        }
-      }}
-    >
-      <DropdownMenuTrigger asChild>
-        <button
-          title={label}
-          className="inline-flex aspect-square h-8 items-center justify-center rounded border border-solid border-f1-border bg-f1-background-inverse-secondary px-0 text-f1-foreground hover:border-f1-border-hover"
-        >
-          <Icon icon={Messages} size="md" />
-        </button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="start" className="max-h-[600px] max-w-md">
-        {updates === null && "loading"}
-        {updates !== null && updates.length > 0 && (
-          <>
-            <a
-              href={updatesPageUrl}
-              className="flex items-center justify-between gap-4 px-4 py-3 text-f1-foreground no-underline visited:text-f1-foreground hover:text-f1-foreground"
-            >
-              <h2 className="text-base font-medium">{label}</h2>
-              <Button
-                variant="outline"
-                round
-                size="sm"
-                icon={ChevronRight}
-                label={label}
-                hideLabel
-              />
-            </a>
-            <FeaturedDropdownItem {...featuredUpdate} />
-            {updates.length > 1 && (
-              <>
-                <DropdownMenuSeparator />
-                {restUpdates.map((update, index) => (
-                  <DropdownItem key={index} {...update} />
-                ))}
-              </>
-            )}
-          </>
-        )}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  )
-}
-
-const FeaturedDropdownItem = ({
-  title,
-  href,
-  imageURL,
-  unread,
-  updated,
-}: ProductUpdate) => {
-  const itemClass = "flex flex-col items-stretch gap-3 w-full"
-  return (
-    <DropdownMenuItem asChild className={itemClass}>
-      <Link
-        href={href}
-        target="_blank"
-        referrerPolicy="no-referrer"
-        rel="noopener noreferrer"
-        className={cn(
-          itemClass,
-          "text-f1-foreground no-underline hover:cursor-pointer"
-        )}
-      >
-        <div className="overflow-clip rounded border border-solid border-f1-border-secondary">
-          <Image
-            src={imageURL}
-            className="block aspect-video w-full object-contain object-center"
-          />
-        </div>
-        <div className="flex items-start gap-4">
-          <div className="flex-1 *:text-base">
-            <h3 className="font-medium">{title}</h3>
-            <p className="font-normal text-f1-foreground-secondary">
-              {updated}
-            </p>
-          </div>
-          {unread && (
-            <div className="mt-1.5 size-2 rounded bg-f1-background-selected-bold" />
-          )}
-        </div>
-      </Link>
-    </DropdownMenuItem>
-  )
-}
-
-const DropdownItem = ({
-  title,
-  href,
-  updated,
-  unread = false,
-}: ProductUpdate) => {
-  const itemClass = cn(
-    "flex flex-col items-stretch gap-3 w-full"
-    // item.critical && "text-f1-foreground-critical"
-  )
-
-  return (
-    <DropdownMenuItem asChild className={itemClass}>
-      <Link
-        href={href}
-        target="_blank"
-        referrerPolicy="no-referrer"
-        rel="noopener noreferrer"
-        className={cn(
-          itemClass,
-          "text-f1-foreground no-underline hover:cursor-pointer"
-        )}
-      >
-        <div className="flex items-start gap-4">
-          <div className="flex-1 *:text-base">
-            <h3 className="font-medium">{title}</h3>
-            <p className="font-normal text-f1-foreground-secondary">
-              {updated}
-            </p>
-          </div>
-          {unread && (
-            <div className="mt-1.5 size-2 rounded bg-f1-background-selected-bold" />
-          )}
-        </div>
-      </Link>
-    </DropdownMenuItem>
   )
 }
